requirements.txt
python-whois==0.9.6
dnspython==2.8.0
PySocks>=1.7.1
PyGithub>=2.1.1
knock-subdomains @ git+https://github.com/guelfoweb/knock.git
python-dotenv>=1.0.0
tldextract>=5.1.0
python-Wappalyzer>=0.3.0
setuptools>=65.0.0  # Required for python-Wappalyzer (pkg_resources)
